+++ CHANGELOG.md
@@ -10,6 +10,19 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [0.3.0] - 2024-08-12
+
+### Changed
+
+- **[BC]** `Run()` now accepts a list of functions, instead of options.
+- **[BC]** All functions now effectively have an "unbounded" outbox size. This
+  means that a send operation to the outbox never blocks once all functions have
+  called `Ready()`.
+
+### Removed
+
+- **[BC]** Removed `Option`, `WithFunc()` and `WithInboxSize()`.
+
 ## [0.2.1] - 2024-08-01
 
 ### Added
@@ -49,6 +62,7 @@ functions executed by `Run()` as "functions".
 [0.1.1]: https://github.com/dogmatiq/minibus/releases/tag/v0.1.1
 [0.2.0]: https://github.com/dogmatiq/minibus/releases/tag/v0.2.0
 [0.2.1]: https://github.com/dogmatiq/minibus/releases/tag/v0.2.1
+[0.3.0]: https://github.com/dogmatiq/minibus/releases/tag/v0.3.0
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD